requirements.txt
streamlit
pandas
plotly
numpy
googlemaps
geocoder
jamaibase
requests
python-dotenv
